+++ crates/nostr-sdk/src/lib.rs
@@ -17,7 +17,7 @@ pub use async_utility;
 #[doc(hidden)]
 pub use nostr::{self, *};
 #[doc(hidden)]
-pub use nostr_database::{self as database, NostrDatabase, NostrDatabaseExt, Profile};
+pub use nostr_database::{self as database, NostrDatabase, NostrDatabaseExt};
 #[doc(hidden)]
 #[cfg(all(target_arch = "wasm32", feature = "indexeddb"))]
 pub use nostr_indexeddb::{IndexedDBError, WebDatabase};