+++ crates/nostr/src/types/mod.rs
@@ -8,6 +8,7 @@ pub mod contact;
 pub mod filter;
 pub mod image;
 pub mod metadata;
+pub mod profile;
 pub mod time;
 pub mod url;
 
@@ -15,5 +16,6 @@ pub use self::contact::Contact;
 pub use self::filter::{Alphabet, Filter, SingleLetterTag};
 pub use self::image::ImageDimensions;
 pub use self::metadata::Metadata;
+pub use self::profile::Profile;
 pub use self::time::Timestamp;
 pub use self::url::{TryIntoUrl, UncheckedUrl, Url};